+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `#[use_modtype]` attribute macro.
 - Trait specific attributes.
+- `modtype::Pow` for `{u64, u128, i8, i16, i32, i64, i128, isize}`
 
 ### Changed
 
@@ -25,7 +26,7 @@
 
 ### Fixed
 
-- `modtype::Pow_{u8, u16, u32, usize}`.
+- `modtype::Pow`.
 
 ## [0.3.0] - 2019-06-22Z